write_req: Close file after writing
When implementing advanced handlers, it is necessary
to know when an upload is done, so further processing
can start.

The AsyncWrite trait has 3 methods:
- poll_write()
- poll_flush()
- poll_close()

Before this patch, poll_write() was called one
or more times, so the writer didn't know when a
transfer was complete. poll_flush() and poll_close()
was never called.

This patch flushes and closes the writer upon completion,
causing both poll_flush() and poll_close() to be called
on the writer.

Signed-off-by: Asbjørn Sloth Tønnesen <ast@fiberby.net>